Adjustments in livebook for running from the Pi
