+++ .gitignore
@@ -21,11 +21,8 @@ erl_crash.dump
 # Temporary files, for example, from tests.
 /tmp/
 /cache
-/priv/libnx_hailo.so
-
-/priv/*.json
-/priv/*.hef
+# Downloaded models, generated labels, and compiled NIF artifacts
+/priv/
 hailort.log
-/priv/*.marker
 .envrc
 .env
