--- nx_hailo/priv/libnx_hailo.so
@@ -1,1 +1,0 @@
-../cache/libnx_hailo.so
