--- priv/.hailo_target
@@ -1,1 +1,0 @@
-hailo10
